Update: Several rule sets
Allow several rule sets to be passed to the DNA gillespie
class. The idea is that interactions in the nucleus are
assumed to happen frequently. However, this is a violation
of the Gillespie algorithm. The idea is that several gillespie
chemical reactions can run in parallel. This is represented
by several rule sets. E.g. there can be one for random association,
one for elongation, and one for DNA repair.