Bugfix: No UV matching sequencing data
The dynamics have been updated and the procedures have been
engineered to match the ChIP seq data. Elongation does not
need to be defined explicitly. This has been incorporated
as an update step. The idea is that the number of necessary
rules can be increased. Elongation are implicitly updated
according to the longest sampled time. DNA and proteins
are also correctly selected after issue with correct
determination DNA segment when none is passed.

The dynamics after radiation haven't been updated so far
and are not working