Bugfix: Simulations of nouV and early NER
The segements variable was not passed as a copy so kept being
changed and included the lesion definition. This is fixed now and
experiments can be repeated. However, damage repair isn't able to
follow a particular program with enzymatic steps with the random
association/dissociation setup. The following implementation updates
will try to interpret DNA lesions as particles that can go through
different steps