Bugfix: Single-cell scale
Problem was that rules for repair dynamics relied on single-cell
scale. However, implementation allowed several associated proteins of
the same kind at the same position. This led to a malfunctioning
repair system. This has been changed. NoUV conditions work
as expected. DNA repair is not fully working yet as the interaction
and the elongation needs to be adapted for the CPD case