Bigfix: Conditional chromatin remains unchanged
Sometimes it's necessary to set conditions which proteins are
already associated to the DNA structure. In the previous
implementation, wherever the conditions were true for the
reactants, the counters where randomly decreased and for
the products inceased. This led to indesired behaviour
for example for stalled Pol2. This is now fixed.
Conditions that need to met are usually given in the reactants
and the products list. Therefore, the recent implementation
keeps track of the positions where protein:DNA counters
were decreased. If the same protein:DNA interaction is defined
in the products list, the same position is increased again.