Bugfix: Handling of lesion update
Lesion updates were previously done while also updating the
counters for the proteins. However, then it was difficult to find
the matching position if the position was supposed to remain the
same. Moreover, this failed if the rule did not include already
associated proteins as reactants.
This has been fixed. Lesion states are updated at
the end of the routine. Positions are determined based on the
previous state.