Update: Export rules
Rules are exported to separate file. Rules are only loaded
from that file. Moreover, added improvements for parameter
variables and flags. The plotting function doesn't need to be
commented out anymore. Parameters are set in the beginning of the
main routine.
Active Pol2 was moved to the Protein class. The string is now
referenced from there.