Bigfix: Proteins remaining on lesion
If the requirement is that proteins remain on a lesion but other
proteins should bind to the very same protein, the update might
file because there's no slot anymore which fulfills the rule
requirements. So if a protein associates to an aleady associated
protein, the procedure now makes sure that it really associates
to it.